pkgs/top-level: add assert to prevent passing elaborated systems
The commit to make package sets composable depends on the fact that
localSystem and crossSystem can be passed forward to lower package sets
for composition. Once we pass a fully elaborated system, this will break
down - getters like "isStatic", "isMusl" etc. will not change with the
package set anymore, but be stuck on the value passed in via those
options.

This is a limitation of lib.systems.elaborate primarily, because it
can't deal with arbitrary overrides properly.

freshBootstrapTools on darwin used to do this, although in this case it
didn't do any harm: There shouldn't be any package sets composed during
bootstrap anyway. Refactor it, to avoid throwing the assert.