zathura: correctly read global configuration file at /etc/zathurarc
zathura made the sysconfdir, i.e. /etc, configurable in 0.5.10. The
default value of sysconfdir unfortunately uses prefix, so zathura would
search for the global config file at $out/etc/zathurarc where it can't
be created (without overriding the derivation):
https://github.com/pwmt/zathura/commit/213a2be033f5638cd085ff50c26ec46d96b089ae

All documentation on zathura claims that /etc/zathurarc is read (e.g.
the man pages installed with the package), so we should actually make
sure that the package behaves as advertised and restore the behavior of
pre 0.5.10.